fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! Fixes #23780: Implement list command
Fixes #23780: Implement list command
